handling multiple error types
